requirements.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.2
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
